refactor replay helpers
